requirements.txt
pelican
markdown
ghp-import
google-genai
python-dotenv
pelican-sitemap
beautifulsoup4
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
Pillow
requests
